requirements-packaging.txt
fastapi
psycopg2-binary
pyinstaller
python-dotenv
sentry-sdk[fastapi]
uvicorn
